Cleaned up core, removed unused classes, moved some classes
